Initial commit to branch 3.x, ver 3.0.0
 On branch 3.x
 Changes to be committed:
	modified:   src/AlgebraCompiler.cpp
	modified:   src/CDatabase.cpp
	modified:   src/CDatabase.h
	modified:   src/CDataset.cpp
	modified:   src/CDataset.h
	modified:   src/CDeviceContext.cpp
	modified:   src/CGrid.cpp
	modified:   src/CPlotDisplay.cpp
	modified:   src/CPlotDisplay.h
	modified:   src/CPlotList.cpp
	modified:   src/CPlotList.h
	modified:   src/CPlotView.cpp
	modified:   src/CPlotView.h
	modified:   src/CPlotWindow.cpp
	modified:   src/CPolarGrid.cpp
	modified:   src/CTransform.cpp
	modified:   src/CTransform.h
	modified:   src/CXyFile.cpp
	modified:   src/CXyFile.h
	modified:   src/CXyPlot.cpp
	modified:   src/CpsDC.cpp
	modified:   src/Makefile
	modified:   src/xyplot.cpp
	modified:   src/xyplot.h

+++ src/CDatabase.h
@@ -2,7 +2,7 @@
 //
 // class CDatabase
 //
-// Copyright 1995-2018 Krishna Myneni
+// Copyright 1995-2023 Krishna Myneni
 // <krishna.myneni@ccreweb.org>
 //
 // This software is provided under the terms of the 
@@ -37,7 +37,7 @@ public:
   CDataset* FindInList (char*);
   int MakeListOfNames (char**);
   void Clear();           // clear the database
-  CDataset* MakeDataset (vector<float*>, int, int, int, char*, char*);
+  CDataset* MakeDataset (vector<double*>, int, int, int, char*, char*);
   CDataset* MakeDataset (CWorkspace41*, int);
   CDataset* MakeDataset (int*);
   int GetColumnCount (char*);
@@ -51,10 +51,5 @@ public:
 };
 
 #endif
-
-
-
-
-
