+++ src/CPlotView.h
@@ -2,7 +2,7 @@
 //
 // Header file for class CPlotView
 //
-// Copyright 2020 Krishna Myneni
+// Copyright 2020--2023 Krishna Myneni
 //
 // This software is provided under the terms of the
 // GNU Affero General Public License (AGPL) v 3.0 or later.
@@ -17,10 +17,10 @@ public:
 	CTransform* m_pCt;
 	CGrid* m_pGrid;
 
-	CPlotView (COORDINATE_SYSTEM, vector<float>);
+	CPlotView (COORDINATE_SYSTEM, vector<double>);
 	~CPlotView ();
-	void SetExtrema (vector<float>);
-	vector<float> GetExtrema ();
+	void SetExtrema (vector<double>);
+	vector<double> GetExtrema ();
 };
 
 #endif
